+++ Tests/DependencyInjection/Fixtures/php/http_client_caching.php
@@ -8,7 +8,6 @@ $container->loadFromExtension('framework', [
                 'cache_pool' => 'foo',
                 'shared' => false,
                 'max_ttl' => 2,
-                'ttl_buffer' => 200,
             ],
         ],
         'scoped_clients' => [